removendo bugs matlab e debugando código
